Revert "build: Optimize is part of the single JAR (#25204)"
This reverts commit 995c23867c6910a69f2b84fd6e925596ade66cf0, reversing
changes made to 7c1d7423c86080cd9348b7543ba5f6d99670277a.

+++ optimize/util/optimize-commons/src/main/java/io/camunda/optimize/service/util/BackoffCalculator.java
@@ -44,9 +44,8 @@ public class BackoffCalculator {
 
   public long calculateSleepTime() {
     currentTimeToWait =
-        Math.min(Math.round((double) currentTimeToWait * BACKOFF_MULTIPLIER), maximumBackoff);
-    nextRetryTime =
-        OffsetDateTime.now().plus(Math.round((double) currentTimeToWait), ChronoUnit.MILLIS);
+        Math.min(Math.round(currentTimeToWait * BACKOFF_MULTIPLIER), maximumBackoff);
+    nextRetryTime = OffsetDateTime.now().plus(Math.round(currentTimeToWait), ChronoUnit.MILLIS);
     return currentTimeToWait;
   }
